LPS-58992 Make XStreamAlias class public
